+++ Functions/TestDrive.ps1
@@ -47,7 +47,8 @@ function Clear-TestDrive ([String[]]$Exclude) {
 function New-RandomTempDirectory {
     do
     {
-        $Path = & $SafeCommands['Join-Path'] -Path $env:TEMP -ChildPath ([Guid]::NewGuid())
+        $tempPath = Get-TempDirectory
+        $Path = & $SafeCommands['Join-Path'] -Path $tempPath -ChildPath ([Guid]::NewGuid())
     } until (-not (& $SafeCommands['Test-Path'] -Path $Path ))
 
     & $SafeCommands['New-Item'] -ItemType Container -Path $Path