+++ CHANGELOG.md
@@ -8,6 +8,15 @@
   -Missing: Different color profiles based on host
   -Missing: Latest features from v3
 
+## 3.4.6 (January 13, 2017)
+  - Fix bug with -Show parameter on PowerShell v4 and older [GH-677]
+  - Add commas to test run summary [GH-676]
+
+## 3.4.5 (January 12, 2017)
+  - Add -Show parameter to filter on-screen output [GH-647]
+  - Add BeIn assertion to assert an item is part of an array [GH-646]
+  - Fix test drive to work on PSCore [GH-643]
+ 
 ## 3.4.4 (November 12, 2016)
   - Add New-MockObject function that creates empty objects for almost any .NET class [GH-635]